Update music licensing information
Update music licensing information, and also remove Scheming Weasel due to weird copyright issue. Basically someone uploaded this music without attribution, as if it was made by him/her. And Twitch decided to assume that the music is not owned by Kevin MacLeod(which is incorrect), and using the music WILL mute the corresponding section in the VOD.